Se hacen cambios de menu
