+++ app/src/main/java/com/cs2/veterinaria/app/adapters/inputs/AdminInput.java
@@ -1,14 +1,16 @@
 package com.cs2.veterinaria.app.adapters.inputs;
 
-import com.cs2.veterinaria.app.domains.model.Person;
+
 import com.cs2.veterinaria.app.domains.model.User;
 import com.cs2.veterinaria.app.domains.services.AdminServices;
 import com.cs2.veterinaria.app.ports.InputPort;
 
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Component;
-import java.util.List;
+
 
 
 import com.cs2.veterinaria.app.adapters.inputs.utils.PersonValidator;
@@ -35,14 +37,11 @@ public class AdminInput implements InputPort {
     private LoginInput loginInput;
 
     private final String MENU = "Ingrese la opcion:"
-            + " \n 1. para crear Usuario."
-            + " \n 2. eliminar usuario" 
+            + " \n 1. Crear Usuario."
+            + " \n 2. Eliminar usuario" 
             + " \n 3. Actualizar datos de usuario"
-            + " \n 4. Registrar Dueño de mascota"
-            + " \n 5. Actualizar mascota de dueño"
-            + " \n 6. Listar usuarios"
-            + " \n 7. Listar dueños"
-            + " \n 8. Salir";
+            + " \n 4. Listar usuarios"            
+            + " \n 5. Salir";
 
     public void menu() throws Exception {
         String option ="0";
@@ -76,37 +75,13 @@ public class AdminInput implements InputPort {
             }
             case "4": {
                 try {
-                    registerOwner();
-                } catch (Exception e) {
-                    System.out.println(e.getMessage());
-                }
-                break;
-            }
-            case "5": {
-                try {
-                    deleteOwner();
-                } catch (Exception e) {
-                    System.out.println(e.getMessage());
-                }
-                break;
-            }
-            case "6": {
-                try {
                     listUsers();
                 } catch (Exception e) {
                     System.out.println(e.getMessage());
                 }
                 break;
             }
-            case "7": {
-                try {
-                    listOwners();
-                } catch (Exception e) {
-                    System.out.println(e.getMessage());
-                }
-                break;
-            }
-            case "8": {
+            case "5": {
                 System.out.println("Hasta una proxima ocación");
                 loginInput.menu();
     break;
@@ -191,37 +166,6 @@ public class AdminInput implements InputPort {
         }
     }
 
-
-    private void registerOwner() {
-        try {
-            System.out.println("Ingrese el nombre del dueño:");
-            String name = personValidator.nameValidator(Utils.getReader().nextLine());
-
-            System.out.println("Ingrese el documento del dueño:");
-            Long document = personValidator.documentValidator(Utils.getReader().nextLine());
-
-            Person person = new Person();
-            person.setName(name);
-            person.setDocument(document);
-
-            adminService.registerOwner(person);
-            System.out.println("Dueño registrado exitosamente.");
-            } catch (Exception e) {
-                System.out.println("Error al registrar el dueño: " + e.getMessage());
-                }
-    }
-
-    private void deleteOwner() {
-        try {
-            System.out.println("Ingrese el documento del dueño a eliminar:");
-            Long document = personValidator.documentValidator(Utils.getReader().nextLine());
-            adminService.deleteOwner(document);
-            System.out.println("Dueño eliminado exitosamente.");
-        } catch (Exception e) {
-            System.out.println("Error al eliminar el dueño: " + e.getMessage());
-            }
-    }
-
     private void listUsers() {
         try {
             System.out.println("Lista de usuarios:");
@@ -233,18 +177,7 @@ public class AdminInput implements InputPort {
             System.out.println("Error al listar los usuarios: " + e.getMessage());
         }
     }
-
-    private void listOwners() {
-        try {
-            System.out.println("Lista de dueños:");
-            List<Person> owners = adminService.listOwners();
-            for (Person owner : owners) {
-                System.out.println("Documento: " + owner.getDocument() + ", Nombre: " + owner.getName());
-            }
-        } catch (Exception e) {
-            System.out.println("Error al listar los dueños: " + e.getMessage());
-        }
-    }
+            
 
 }
