+++ app/src/main/java/com/cs2/veterinaria/app/domains/services/veterinarianService.java
@@ -6,8 +6,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import com.cs2.veterinaria.app.adapters.pet.entity.PetEntity;
 import com.cs2.veterinaria.app.domains.model.HistoryClinic;
 import com.cs2.veterinaria.app.domains.model.Order;
+import com.cs2.veterinaria.app.domains.model.Person;
 import com.cs2.veterinaria.app.ports.ClinicalPort;
 import com.cs2.veterinaria.app.ports.OrderPort;
+import com.cs2.veterinaria.app.ports.PersonPort;
+import com.cs2.veterinaria.app.ports.PetOwnerPort;
 import com.cs2.veterinaria.app.ports.UserPort;
 
 import java.util.List;
@@ -20,6 +23,12 @@ import lombok.Setter;
 @NoArgsConstructor
 @Service
 public class VeterinarianService {
+
+    @Autowired
+    private PersonPort personPort;
+
+    @Autowired
+    private PetOwnerPort petOwnerPort;
 
     @Autowired
     private ClinicalPort clinicalPort;
@@ -62,5 +71,30 @@ public class VeterinarianService {
     // Método para consultar todas las órdenes de medicamentos
     public List<Order> getAllOrders() {
         return orderPort.findAllOrders();
+    }
+
+    public void registerOwner(Person person) throws Exception {
+        // Verificar si la persona ya existe
+        if (personPort.existPerson(person.getDocument())) {
+            throw new Exception("Ya existe una persona con ese documento");
+        }
+    
+        // Guardar la persona en la base de datos
+        personPort.savePerson(person);
+    
+        // Verificar que el person_id se haya asignado correctamente
+        if (person.getPersonId() == 0) {
+            throw new Exception("Error al asignar el ID de la persona.");
+        }
+    
+        // Guardar el dueño de la mascota
+        petOwnerPort.savePetOwner(person);
+    }
+
+    public void deleteOwner(Long document) throws Exception {
+        if (!personPort.existPerson(document)) {
+            throw new Exception("No existe una persona con ese documento");
+        }
+        personPort.deletePerson(document);
     }
 }
